requirements.txt
pyttsx3
pywhatkit
datetime
wikipedia
pyjokes
geopy
python-newsapi
requests
pyautogui
clipboard
pyaudio
